+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.0.2 (2019-XX-XX)
 
 * Fixed #6: `workbench` does not crash anymore when certain other plugins are loaded. Background: Their UI exposes its Qt parent as a property instead of a method.
+* Fixed #7: `workbench` can start even of other plugins contain unnamed UI elements.
 
 ## 0.0.1 (2019-09-01)